extract piece
