+++ rust/src/hexchess/mod.rs
@@ -1,6 +1,7 @@
 pub mod bitboard;
 pub mod game;
 pub mod hexchess;
+pub mod piece;
 pub mod pieces;
 pub mod position;
 pub mod san;
