+++ rust/src/lib.rs
@@ -2,7 +2,8 @@ pub mod constants;
 pub mod hexchess;
 pub mod macros;
 
-pub use constants::{Color, Piece};
+pub use constants::Color;
 pub use hexchess::hexchess::Hexchess;
+pub use hexchess::piece::Piece;
 pub use hexchess::san::San;
 pub use hexchess::utils::{fen_index, position};
